requirements.txt
pytube
moviepy
openai-whisper
deep-translator
requests
streamlit
